% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_max.R
\name{f.which.max}
\alias{f.which.max}
\title{Change behavior of which.max to work as if it had an na.rm=TRUE option}
\usage{
f.which.max(i)
}
\arguments{
\item{i}{A vector}
}
\description{
Change behavior of which.max to work as if it had an na.rm=TRUE option using
https://gis.stackexchange.com/questions/392666/app-function-returns-a-wrong-output-format
answer recommended by R Hijmans.
}
\examples{
\dontrun{
f.which.max(1:5)
f.which.max(NA)
}
}
