% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdDates.R
\name{msdDates}
\alias{msdDates}
\title{Pulling Dates Relevant to the MSD Calculations}
\usage{
msdDates(x, peakwindow1, minwindow1, minwindow2, secondEndDate)
}
\arguments{
\item{x}{Date vector}

\item{peakwindow1}{desired date in MMDD format to begin search for early peak}

\item{minwindow1}{desired date in MMDD format to begin search for minimum}

\item{minwindow2 }{desired date in MMDD format to end search for minimum}

\item{peakwindow2}{desired date in MMDD format to end search for late peak}
}
\value{
Date vector containing each year's beginning date, end date, and critical MSD dates
}
\description{
This function pulls the dates that correspond with the Mid Summer Drought, and is used in conjuction with the MSD function.
It also pulls the first and last day of each calendar year provided.
}
\examples{
x <- seq(from = as.Date("1981-01-01"), to = as.Date("1985-12-31"), by = "day")
output <- msdDates(x, peakwindow1 = "05-01", minwindow1 = "06-01", minwindow2 = "08-31", peakwindow2 = "10-31")

}
