% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdFinal.R
\name{msdFinal}
\alias{msdFinal}
\title{Mid Summer Drought Function}
\usage{
msdFinal(x, firstStartDate, firstEndDate, secondStartDate, secondEndDate, quantity, window)
}
\arguments{
\item{x}{Date vector}

\item{firstStartDate}{desired date in MMDD format to begin analysis (window 1)}

\item{firstEndDate}{desired date in MMDD format to end analysis (window 1)}

\item{secondStartDate}{desired date in MMDD format to begin analysis (window 2)}

\item{secondEndDate}{desired date in MMDD format to end analysis (window 2)}

\item{quantity}{Size of filter}
}
\value{
Dataframe of all relevant MSD Statistics
}
\description{
Generates the relevant statistics for the Mid Summer Drought

HERE
}
\examples{
# using timeseries (ts)
# output = msdFinal(ts, firstStartDate="05-01", firstEndDate="08-31",
# secondStartDate ="06-01", secondEndDate="10-31", quantity = 2, window = 31)

}
