% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdGraph.R
\name{msdGraph}
\alias{msdGraph}
\title{Mid Summer Drought Time Series Graphs}
\usage{
msdGraph(x, year, firstStartDate, firstEndDate,
secondStartDate, secondEndDate, quantity,
window, time)
}
\arguments{
\item{x}{Vector or TimeSeries}

\item{year}{Year of interest}

\item{firstStartDate}{desired date in MMDD format to begin analysis (window 1)}

\item{firstEndDate}{desired date in MMDD format to end analysis (window 1)}

\item{secondStartDate}{desired date in MMDD format to begin analysis (window 2)}

\item{secondEndDate}{desired date in MMDD format to end analysis (window 2)}

\item{quantity}{Amount of times the filter is run}

\item{window}{Size of filter}

\item{time}{Vector of dates (not needed for TimeSeries inputs)}
}
\value{
Graph of Time Series Data
}
\description{
Plots the Time Series of Mid Summer Drought data

The input must be in the form of daily data, with the first data point being January 1st of a respective year.
}
\examples{
# graph = msdGraph(x, year, firstStartDate="05-01", firstEndDate="08-31",
# secondStartDate ="06-01", secondEndDate="10-31", quantity = 2, window = 31)

}
