% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdMain.R
\name{msdMain}
\alias{msdMain}
\title{Mid Summer Drought Function}
\usage{
msdMain(x, firstStartDate, firstEndDate,
secondStartDate, secondEndDate, quantity,
window, time)
}
\arguments{
\item{x}{Vector or TimeSeries}

\item{peakwindow1}{desired date in MMDD format to begin analysis (window 1)}

\item{minwindow1}{desired date in MMDD format to end analysis (window 1)}

\item{minwindow2}{desired date in MMDD format to begin analysis (window 2)}

\item{peakwindow2}{desired date in MMDD format to end analysis (window 2)}

\item{quantity}{amount of times the filter is run}

\item{window}{size of filter}

\item{timeVector}{vector of dates (not needed for xts inputs)}
}
\value{
Data frame of all relevant MSD Statistics
}
\description{
Generates all relevant statistics for the Mid Summer Drought
}
\examples{
# using timeseries (ts)
# output = msdMain(ts, peakwindow1="05-01", minwindow1="08-31",
# minwindow2 ="06-01", peakwindow2="10-31", quantity = 2, window = 31)

}
