% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdStats.R
\name{msdStats}
\alias{msdStats}
\title{Main Mid Summer Drought Calculation Function}
\usage{
msdStats(x, dates, fcn)
}
\arguments{
\item{x}{RasterBrick or TimeSeries}

\item{quantity}{Amount of times the data should be filtered. The default option is 2}

\item{dates}{Vector of Dates (from the msdDates function)}

\item{fcn}{Specify what values to be pulled from the function. Options are 'duration', 'intensity', 'firstMax', 'secondMax', 'min', and 'mindex'.}
}
\value{
SpatRaster or TimeSeries of Yearly data
}
\description{
This function calculates the different statistics of the mid summer drought from a RasterBrick or a Time Series.

The input must be in the form of daily data, with the first data point being January 1st of a respective year.
If x is a SpatRaster, then the output is a SpatRaster with a data point for each year.
}
\examples{
# using timeseries
# r <- apply(timeseries, MARGIN = 2, FUN = msdStats, quantity = 2, dates = keyDatesTS, fcn="secondMax")

}
