% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R, R/timeseries.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Steps to produce the raster sample data provided in the MSD package}
\format{
## `see data-raw file on GitHub to see process to generating timeseries and raster data`

## `see data-raw file on GitHub to see process to generating timeseries and raster data`
}
\source{
University of California, Santa Barbara Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS)

University of California, Santa Barbara Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS)
}
\usage{
timeseries

timeseries
}
\description{
Daily precipitation data recorded from 1981 to 1985, trimmed to only highlight Nicaragua

Daily precipitation data recorded from 1981 to 1985, trimmed to only highlight Nicaragua
}
\keyword{datasets}
