% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdDates.R
\name{msdDates}
\alias{msdDates}
\title{Pulling Dates Relevant to the MSD Calculations}
\usage{
msdDates(x, firstStartDate, firstEndDate, secondStartDate, secondEndDate)
}
\arguments{
\item{x}{vector of dates}

\item{firstStartDate}{desired date in MMDD format to begin analysis (window 1)}

\item{firstEndDate}{desired date in MMDD format to end analysis (window 1)}

\item{secondStartDate}{desired date in MMDD format to begin analysis (window 2)}

\item{secondEndDate}{desired date in MMDD format to end analysis (window 2)}
}
\value{
Date vector containing each year's beginning date, end date, and critical MSD dates
}
\description{
This function pulls the dates that correspond with the Mid Summer Drought, and is used in conjuction with the MSD function.
It also pulls the first and last day of each calendar year provided.
}
\examples{
x <- seq(from = as.Date("1981-01-01"), to = as.Date("1985-12-31"), by = "day")
msdDates(x, firstStartDate="05-01",firstEndDate="08-31",
secondStartDate ="06-01",secondEndDate="10-31")

}
