% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdMain.R
\name{msdMain}
\alias{msdMain}
\title{Mid Summer Drought Function}
\usage{
msdMain(x, peakwindow1, minwindow1,
minwindow2, peakwindow2, quantity,
window, timeVector)
}
\arguments{
\item{x}{xts or vector of data}

\item{peakwindow1}{desired date in MMDD format to begin analysis (window 1)}

\item{minwindow1}{desired date in MMDD format to end analysis (window 1)}

\item{minwindow2}{desired date in MMDD format to begin analysis (window 2)}

\item{peakwindow2}{desired date in MMDD format to end analysis (window 2)}

\item{quantity}{amount of times the filter is run}

\item{window}{size of filter}

\item{timeVector}{vector of dates (not needed for xts inputs)}
}
\value{
Data frame of all relevant MSD Statistics
}
\description{
Generates all relevant statistics for the Mid Summer Drought by running the msdStats function for every
applicable metric. The output of msdMain is a dataframe containing every msdStats output for the available years of data.
}
\examples{

data("timeseries")
ts <- timeseries
df <- msdrought::msdMain(ts)

}
