% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Timeseries of precipitation data}
\format{
## `timeseries`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2, iso3}{2 & 3 letter ISO country codes}
  \item{year}{Year}
  ...
}
}
\source{
University of California, Santa Barbara Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS)
}
\usage{
timeseries
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
