% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Steps to produce the sample data provided in the MSD package}
\format{
## `see data-raw file on GitHub to see process to generating timeseries and raster data`
Daily precipitation data recorded from 1981 to 1985, trimmed to only highlight Nicaragua
}
\source{
University of California, Santa Barbara Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS)
}
\usage{
timeseries
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
