% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{msdrought package included timeseries data}
\format{
## `timeseries`
An xts object with precipitation data over a series of five years:
\describe{
  \item{Date}{date in YYYY-MM-DD format}
  \item{Precipitation}{daily precipitation in millimeters (mm)}
  ...
}
}
\source{
<https://www.chc.ucsb.edu/data/chirps>
}
\usage{
timeseries
}
\description{
A subset of data from the UCSB CHIRPS data set
}
\keyword{datasets}
